requirements-dev.txt
pytest-django
coverage
